+++ mapper/pom.xml
@@ -20,11 +20,6 @@
       <artifactId>jakarta.persistence-api</artifactId>
     </dependency>
     <dependency>
-      <groupId>com.google.code.findbugs</groupId>
-      <artifactId>annotations</artifactId>
-      <version>${findbugs.version}</version>
-    </dependency>
-    <dependency>
       <groupId>io.smallrye</groupId>
       <artifactId>jandex</artifactId>
       <version>${jandex.version}</version>
