+++ pom.xml
@@ -47,17 +47,18 @@
     <!-- Plugins -->
     <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
     <maven-surefire-plugin.version>3.5.2</maven-surefire-plugin.version>
-    <spotless-maven-plugin.version>2.43.0</spotless-maven-plugin.version>
-    <maven-source-plugin.version>3.2.0</maven-source-plugin.version>
-    <maven-javadoc-plugin.version>3.3.1</maven-javadoc-plugin.version>
-    <maven-site-plugin.version>3.8.2</maven-site-plugin.version>
+    <spotless-maven-plugin.version>2.44.2</spotless-maven-plugin.version>
+    <maven-source-plugin.version>3.3.1</maven-source-plugin.version>
+    <maven-javadoc-plugin.version>3.11.2</maven-javadoc-plugin.version>
+    <maven-site-plugin.version>3.21.0</maven-site-plugin.version>
+    <maven-jar-plugin.version>3.3.0</maven-jar-plugin.version>
+    <build-helper-maven-plugin.version>3.6.0</build-helper-maven-plugin.version>
     <maven.version.ignore>(?i).*(?:-|\.)(alpha|beta|m|rc|cr)(-?\d+)?$</maven.version.ignore>
     <license-maven-plugin.version>4.6</license-maven-plugin.version>
     <maven-gpg-plugin.version>3.0.1</maven-gpg-plugin.version>
     <jandex-maven-plugin.version>3.1.6</jandex-maven-plugin.version>
 
     <spring-boot.version>3.4.2</spring-boot.version>
-    <findbugs.version>3.0.1</findbugs.version>
     <skip.formatting>false</skip.formatting>
     <archunit-junit5.version>1.3.0</archunit-junit5.version>
     <jandex.version>3.2.3</jandex.version>
@@ -195,19 +196,45 @@
       <build>
         <plugins>
           <plugin>
+            <groupId>org.codehaus.mojo</groupId>
+            <artifactId>build-helper-maven-plugin</artifactId>
+            <version>${build-helper-maven-plugin.version}</version>
+            <executions>
+              <execution>
+                <id>regex-property</id>
+                <goals>
+                  <goal>regex-property</goal>
+                </goals>
+                <configuration>
+                  <name>java-module-name</name>
+                  <value>${project.artifactId}</value>
+                  <regex>-</regex>
+                  <replacement>.</replacement>
+                  <failIfNoMatch>false</failIfNoMatch>
+                </configuration>
+              </execution>
+            </executions>
+          </plugin>
+          <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-jar-plugin</artifactId>
+            <version>${maven-jar-plugin.version}</version>
+            <configuration>
+              <archive>
+                <manifest>
+                  <addDefaultImplementationEntries>true</addDefaultImplementationEntries>
+                </manifest>
+                <manifestEntries>
+                  <Automatic-Module-Name>${java-module-name}</Automatic-Module-Name>
+                </manifestEntries>
+              </archive>
+            </configuration>
+          </plugin>
+          <plugin>
             <groupId>org.apache.maven.plugins</groupId>
             <artifactId>maven-source-plugin</artifactId>
             <version>${maven-source-plugin.version}</version>
-            <configuration>
-              <attach>true</attach>
-            </configuration>
             <executions>
-              <execution>
-                <id>package</id>
-                <goals>
-                  <goal>jar-no-fork</goal>
-                </goals>
-              </execution>
               <execution>
                 <id>attach-sources</id>
                 <goals>
