+++ starter/src/main/java/tw/com/softleader/data/jpa/spec/starter/SpecificationResolverCodecBuilder.java
@@ -18,26 +18,27 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package tw.com.softleader.data.jpa.spec.repository.usecase;
+package tw.com.softleader.data.jpa.spec.starter;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import java.util.function.Function;
+import org.springframework.core.Ordered;
+import tw.com.softleader.data.jpa.spec.SpecCodec;
+import tw.com.softleader.data.jpa.spec.SpecificationResolver;
+import tw.com.softleader.data.jpa.spec.starter.autoconfigure.SpecMapperAutoConfiguration;
 
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Entity
-public class Customer {
-
-  @Id @GeneratedValue Long id;
-
-  String name;
-
-  Integer age;
+/**
+ * Represents a builder that accepts {@code SpecCodec} and produces a {@code SpecificationResolver},
+ * for autoconfigure to scan.
+ *
+ * @author Matt Ho
+ * @see SpecCodec
+ * @see SpecificationResolver
+ * @see SpecMapperAutoConfiguration
+ */
+public interface SpecificationResolverCodecBuilder
+    extends Function<SpecCodec, SpecificationResolver>, Ordered {
+  @Override
+  default int getOrder() {
+    return 0;
+  }
 }
