+++ starter/src/main/java/tw/com/softleader/data/jpa/spec/starter/repository/support/JpaRepositoryFactoryBeanPostProcessor.java
@@ -18,7 +18,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package tw.com.softleader.data.jpa.spec.repository.support;
+package tw.com.softleader.data.jpa.spec.starter.repository.support;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
