+++ starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-tw.com.softleader.data.jpa.spec.autoconfigure.SpecMapperAutoConfiguration
+tw.com.softleader.data.jpa.spec.starter.autoconfigure.SpecMapperAutoConfiguration
