+++ starter/src/test/java/tw/com/softleader/data/jpa/spec/starter/repository/CustomizeSkippingStrategyTest.java
@@ -18,7 +18,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package tw.com.softleader.data.jpa.spec.repository;
+package tw.com.softleader.data.jpa.spec.starter.repository;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
